app/requirements.txt
pymongo
python-jose[cryptography]
passlib[bcrypt]
fastapi-login
python-multipart
authlib
